Classify both headline and body sentiment
