+++ backend/app/__init__.py
@@ -1,0 +1,2 @@
+from .main import app
+__all__ = ["app"]
