+++ backend/requirements.txt
@@ -22,7 +22,6 @@ exchange_calendars==4.11
 executing==2.2.0
 fastapi==0.116.1
 fastjsonschema==2.21.2
-feedparser==6.0.11
 flatbuffers==25.2.10
 fonttools==4.59.1
 frozendict==2.4.6
@@ -122,3 +121,8 @@ Werkzeug==3.1.3
 wheel==0.45.1
 wrapt==1.17.3
 yfinance==0.2.65
+
+GoogleNews==1.6.9
+trafilatura==1.6.1
+transformers==4.46.2
+torch==2.5.1
